+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixes
     - Fix modules preventing alert menu from staying open
     - Fix for sometimes attempting to select next event in wrong table
+    - Fix exception if Context Action environment match regex does not match on given environment
 
 # v2.0.0-s5.5.3
 - 'Official Release'